add fire protection too
