+++ src/main/java/net/countercraft/movecraft/worldguard/MovecraftWorldGuard.java
@@ -71,6 +71,7 @@ public final class MovecraftWorldGuard extends JavaPlugin {
         getServer().getPluginManager().registerEvents(new CraftRotateListener(), this);
         getServer().getPluginManager().registerEvents(new CraftSinkListener(), this);
         getServer().getPluginManager().registerEvents(new CraftTranslateListener(), this);
+        getServer().getPluginManager().registerEvents(new FireSpreadListener(), this);
         getServer().getPluginManager().registerEvents(new ExplosionListener(), this);
     }
